Mgmt: add annotate-mgmt pass
