+++ lib/Transforms/OptimizeRelinearization/BUILD
@@ -19,8 +19,8 @@ cc_library(
         "@heir//lib/Analysis/SecretnessAnalysis",
         "@heir//lib/Dialect:Utils",
         "@heir//lib/Dialect/Mgmt/IR:Dialect",
+        "@heir//lib/Dialect/Mgmt/Transforms",
         "@heir//lib/Dialect/Secret/IR:Dialect",
-        "@heir//lib/Transforms/SecretInsertMgmt:SecretInsertMgmtPatterns",
         "@llvm-project//llvm:Support",
         "@llvm-project//mlir:Analysis",
         "@llvm-project//mlir:FuncDialect",
